fix: manifest에서 service 객체 타입 interface to class
