+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,6 @@
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" android:maxSdkVersion="32" />
     <uses-permission android:name="android.permission.READ_MEDIA_IMAGES" />
     <uses-permission android:name="android.permission.READ_MEDIA_VISUAL_USER_SELECTED" />
-    <!--    <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />-->
 
     <application
         android:allowBackup="true"
@@ -44,7 +43,8 @@
         <!-- Required for Android 10 (API level 29) and higher. -->
         <service
             android:name=".userlocation.LocationService"
-            android:foregroundServiceType="location" >
+            android:foregroundServiceType="location"
+            android:exported="false">
         </service>
     </application>
 </manifest>
